Adds code to generate rows and columns.
At this point, the date comes from the date selected by the user at the top
of the form. The members come from a get request already filtered to
include current members only. The custom criteria comes from a get
request of the tracking criteria.

Most of the interface is complete, now needs to create and send the form
data as an activity log entry.