Moves username or date to top when filtered.
To save space on mobile in particular, selecting a username and/or a date
will remove that column from the table and instead add it as a heading above
the table. This also makes the table less busy, since there won't be a
repetitive field that doesn't actually help convey the data.