Starts separation of different activity logs.
Previous version only allowed for one type of activity log to be viewed.
This code has added a separate section for adding in each activity to
track and what is its type of log (daily, summary, etc). Each criteria
is now associated with a particular type of activity.

Still need to add functionality to create a new activity log (can only
RUD at this point, no C). May need to rework the list component to
handle multiple types of lists (in this case, one for activities and one
for criteria) and intersperse the add button where appropriate.